Add nullable to openapi.json & sanitations.md
Update docs/spec/openapi.json file to add nullable to the fields that
are optional in the API response. This ensures no json parsing errors
when the optional fields are not present in the response.

Update docs/spec/sanitations.md file to reflect the changes done on
docs/spec/openapi.json file.