Update and fix test cases for client testing.
Update positive and negative test cases for the client to handle certain
edge cases. This includes tests for invalid app id and payload with
complete settings (i.e. including all optionals).

Fix the test cases to deal with the DB delay happening at HubSpot's end.
HubSpot requires a delay around a minute to update the settings in the
DB. Test cases with method GET fails due to this delay, if it is not
accounted for.