Add setup guide in README.md
Add setup guide section in documentation with screenshots for setting up
the video conferrencing settings module.

This gives a step-by-step easy to follow guide on how to get
authentication tokens from the HubSpot, how to set up the test apps and
developer account in HubSpot in order to use this module.