Remove delay in live tests for get method
Adding a delay in the tests could cause intermittent failures if the
delay is not sufficient for the particular situation. This commit
removes the delay in the live tests for the get method.

This will result in couple of test failures in the live tests but consistent
test results are achieved.